FreeStack safety

+++ alloc/src/size_classes.rs
@@ -3,7 +3,7 @@ pub const NUM_SIZE_CLASSES: usize = 5;
 pub const SIZE_CLASSES: [u32; NUM_SIZE_CLASSES] = [256, 512, 1024, 2048, 4096];
 pub const MIN_SIZE: u32 = SIZE_CLASSES[0];
 pub const MAX_SIZE: u32 = SIZE_CLASSES[NUM_SIZE_CLASSES - 1];
-const BASE_SHIFT: u32 = SIZE_CLASSES[0].trailing_zeros() as u32;
+const BASE_SHIFT: u32 = SIZE_CLASSES[0].trailing_zeros();
 
 const _: () = assert!(
     NUM_SIZE_CLASSES < 256,
